Add RequireDaprApiTokenAttribute for API token validation (#42)
* Add RequireDaprApiTokenAttribute for API token validation

This commit introduces the `RequireDaprApiTokenAttribute` class, which implements the `IAuthorizationFilter` interface. This attribute enforces the presence of a specific API token in request headers for controller actions in an ASP.NET Core application. If the token is missing, empty, or does not match the expected value from the configuration, a 401 Unauthorized response is returned. The class also includes functionality to load the expected token from the application settings.

Resolves #41

* Add capability to remove servers from OpenAPI description file.

#43